+++ ibis/common/dispatch.py
